Ajustando arquivos
