+++ Sistema_Credenciamento.py/User_CRUD.py
@@ -1,17 +1,8 @@
-import mysql.connector
-
-def conectar():
-    
-    return mysql.connector.connect(
-        host="localhost",
-        user="root",
-        password="Projetocdc2025!",
-        database="teste2"
-    )
+import conectar_banco
 
 def cadastrar():
     
-    con = conectar()
+    con = conectar_banco.conectar()
     cursor = con.cursor()
 
     nome = input("\nDigite o nome do usuário: ")
@@ -32,7 +23,7 @@ def cadastrar():
 
 def atualizar_porId():
     
-    con = conectar()
+    con = conectar_banco.conectar()
     cursor = con.cursor()
 
     id_usuario = int(input("\nDigite o ID do usuário que deseja atualizar: "))
@@ -54,7 +45,7 @@ def atualizar_porId():
 
 def deletar_porId():
 
-    con = conectar()
+    con = conectar_banco.conectar()
     cursor = con.cursor()
 
     id_usuario = int(input("\nDigite o ID do usuário que deseja deletar: "))
@@ -69,7 +60,7 @@ def deletar_porId():
 
 def listar():
 
-    con = conectar()
+    con = conectar_banco.conectar()
     cursor = con.cursor()
 
     sql = "SELECT * FROM USUARIO"
@@ -90,48 +81,3 @@ def listar():
 
     cursor.close()
     con.close()
-def criar_banco():
-    con = mysql.connector.connect(
-        host="localhost",
-        user="root",
-        password="Projetocdc2025!",
-        database="teste2"
-    )
-    cursor = con.cursor()
-
-    with open("teste_com_menu/teste2.sql", "r") as arquivo:
-        comandos = arquivo.read().split(';')
-        for comando in comandos:
-            if comando.strip() != "":
-                cursor.execute(comando)
-    con.commit()
-    cursor.close()
-    con.close()
-
-def menu():
-    while True:
-        print("\n--- MENU ---")
-        print("1 - Cadastrar Usuário")
-        print("2 - Atualizar Usuário por ID")
-        print("3 - Deletar usuário por ID")
-        print("4 - Listar usuários ")
-        print("5 - Sair")
-
-        opcao = input("Escolha uma opção: ")
-
-        if opcao == "1":
-            cadastrar()
-        elif opcao == "2":
-            atualizar_porId()
-        elif opcao == "3":
-            deletar_porId()
-        elif opcao == "4":
-            listar()
-        elif opcao == "5":
-            print("Saindo...")
-            break
-        else:
-            print("Opção inválida.")
-
-criar_banco()
-menu()
