+++ Sistema_Credenciamento.py/conectar_banco.py
@@ -1,0 +1,10 @@
+import mysql.connector
+
+def conectar():
+    
+    return mysql.connector.connect(
+        host="localhost",
+        user="root",
+        password="Projetocdc2025!",
+        database="teste2"
+    )
